backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
google-genai
python-dotenv
anomaly